Add permissions support for DeploymentActionsMenu component

+++ widgets/common/src/DeploymentActionsMenu.jsx
@@ -10,19 +10,28 @@ export const actions = Object.freeze({
 
 const translate = (key, params) => Stage.i18n.t(`widgets.common.deployments.actionsMenu.${key}`, params);
 const menuItems = [
-    { name: actions.install, icon: 'play' },
-    { name: actions.update, icon: 'edit' },
-    { name: actions.setSite, icon: 'building' },
-    { name: actions.manageLabels, icon: 'tags' },
-    { name: actions.uninstall, icon: 'recycle' },
-    { name: actions.delete, icon: 'trash alternate' },
-    { name: actions.forceDelete, icon: 'trash' }
-].map(item => ({ ...item, key: item.name, content: translate(item.name) }));
+    { name: actions.install, icon: 'play', permission: 'execution_start' },
+    { name: actions.update, icon: 'edit', permission: 'deployment_update_create' },
+    { name: actions.setSite, icon: 'building', permission: 'deployment_set_site' },
+    { name: actions.manageLabels, icon: 'tags', permission: 'deployment_create' },
+    { name: actions.uninstall, icon: 'recycle', permission: 'execution_start' },
+    { name: actions.delete, icon: 'trash alternate', permission: 'deployment_delete' },
+    { name: actions.forceDelete, icon: 'trash', permission: 'deployment_delete' }
+];
 
-export default function DeploymentActionsMenu({ onActionClick, trigger }) {
+export default function DeploymentActionsMenu({ onActionClick, toolbox, trigger }) {
     const {
-        Basic: { Menu, Popup, PopupMenu }
+        Basic: { Menu, Popup, PopupMenu },
+        Utils: { isUserAuthorized }
     } = Stage;
+
+    const managerState = toolbox.getManagerState();
+    const items = menuItems.map(item => ({
+        ...item,
+        key: item.name,
+        content: translate(item.name),
+        disabled: !isUserAuthorized(item.permission, managerState)
+    }));
 
     function onItemClick(event, { name }) {
         onActionClick(name);
@@ -31,13 +40,14 @@ export default function DeploymentActionsMenu({ onActionClick, trigger }) {
     return (
         <PopupMenu className="deploymentActionsMenu">
             {trigger && <Popup.Trigger>{trigger}</Popup.Trigger>}
-            <Menu pointing vertical onItemClick={onItemClick} items={menuItems} />
+            <Menu pointing vertical onItemClick={onItemClick} items={items} />
         </PopupMenu>
     );
 }
 
 DeploymentActionsMenu.propTypes = {
     onActionClick: PropTypes.func.isRequired,
+    toolbox: Stage.PropTypes.Toolbox.isRequired,
     trigger: PropTypes.node
 };
